signal handler fixes
