+++ cmd/swapd/signal_handler_test.go
